+++ examples/end.php
@@ -12,15 +12,13 @@ $factory->createClient('localhost:11211')->then(
     function (Client $client) {
         $client->set('name', 'test')->then(function(){
             echo "The value was stored\n";
-        }, function(Exception $e){
-            echo 'set: ', $e->getMessage(), "\n";
         });
-        $client->close();
+        $client->end();
         $client->get('name')->then(function($data){
             var_dump($data);
             echo "The value was retrieved\n";
         }, function(Exception $e){
-            echo 'get: ', $e->getMessage(), "\n";
+            echo $e->getMessage(), "\n";
         });
     },
     function(Exception $e){