+++ tests/ClientTest.php
@@ -6,27 +6,28 @@ use Mockery;
 use Mockery\MockInterface;
 use React\Stream\DuplexStreamInterface;
 use seregazhuk\React\Memcached\Client;
+use seregazhuk\React\Memcached\Exception\ConnectionClosedException;
 use seregazhuk\React\Memcached\Exception\Exception;
+use seregazhuk\React\Memcached\Exception\WrongCommandException;
 use seregazhuk\React\Memcached\Protocol\Parser;
 use seregazhuk\React\Memcached\Protocol\Request\Factory as RequestFactory;
 use seregazhuk\React\Memcached\Protocol\Response\Factory as ResponseFactory;
 
-class StreamingClientTest extends TestCase
-{
-    /**
-     * @var DuplexStreamInterface|MockInterface
-     */
+class ClientTest extends PromiseTestCase
+{    /**
+ * @var DuplexStreamInterface|MockInterface
+ */
     protected $stream;
-
-    /**
-     * @var Client
-     */
-    protected $client;
 
     /**
      * @var Parser|MockInterface
      */
     protected $parser;
+
+    /**
+     * @var Client
+     */
+    protected $client;
 
     protected function setUp()
     {
@@ -48,25 +49,20 @@ class StreamingClientTest extends TestCase
     /** @test */
     public function it_rejects_a_promise_when_unsupported_command_is_called()
     {
-        $this->parser->shouldReceive('makeRequest')->andReturn("not_valid\n\r");
-        $this->stream->shouldReceive('write')->once();
         $promise = $this->client->not_valid();
-
-        $this->client->resolveRequests(['not_valid']);
-
-        $this->expectPromiseRejects($promise);
+        $this->expectPromiseRejectsWith($promise, WrongCommandException::class);
     }
 
     /** @test */
     public function it_resolves_a_promise_with_data_from_response()
     {
-        $this->parser->shouldReceive('makeRequest')->andReturn("version\n\r");
+        $this->parser->shouldReceive('makeRequest')->once();
         $this->stream->shouldReceive('write')->once();
         $promise = $this->client->version();
 
-        $this->client->resolveRequests(['version']);
+        $this->client->resolveRequests(['12345']);
 
-        $this->expectPromiseResolves($promise);
+        $this->expectPromiseResolvesWith($promise, '12345');
     }
 
     /** @test */
@@ -79,35 +75,44 @@ class StreamingClientTest extends TestCase
     /** @test */
     public function it_rejects_pending_request_when_closing()
     {
-        $this->parser->shouldReceive('makeRequest')->andReturn("version\n\r");
+        $this->parser->shouldReceive('makeRequest')->once();
         $this->stream->shouldReceive('write')->once();
         $this->stream->shouldReceive('close')->once();
         $promise = $this->client->version();
 
         $this->client->close();
 
-        $this->expectPromiseRejects($promise);
+        $this->expectPromiseRejectsWith($promise, ConnectionClosedException::class);
     }
 
     /** @test */
     public function it_rejects_all_new_requests_when_closed()
     {
-        $this->parser->shouldReceive('makeRequest')->andReturn("version\n\r");
+        $this->parser->shouldNotReceive('makeRequest');
         $this->stream->shouldReceive('close')->once();
 
         $this->client->close();
         $promise = $this->client->version();
-        $this->expectPromiseRejects($promise);
+        $this->expectPromiseRejectsWith($promise, ConnectionClosedException::class);
     }
 
     /** @test */
     public function it_rejects_all_new_requests_when_ending()
     {
-        $this->parser->shouldReceive('makeRequest')->andReturn("version\n\r");
+        $this->parser->shouldNotReceive('makeRequest');
         $this->stream->shouldReceive('close')->once();
 
         $this->client->end();
         $promise = $this->client->version();
-        $this->expectPromiseRejects($promise);
+        $this->expectPromiseRejectsWith($promise, ConnectionClosedException::class);
+    }
+
+	/** @test */
+	public function it_emits_close_event_when_closing()
+	{
+		$this->stream->shouldReceive('close')->once();
+
+		$this->client->on('close', $this->expectCallableOnce());
+		$this->client->end();
     }
 }