+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## v0.1.2 - ?
+### Fixed:
+ - handling broken connections
+ - `createClient()` uses default Memcached address
+ - don't allow new requests when *is ending* or *is closed*
+ - write response resolving
+ - wrong commands are rejected immediately
+ - handle failed commands 
+ - handle broken connection
+ 
+### Added:
+ - client can emit events (`close`, `error`)
+
 ## v0.1.1 - 2017-10-19
 ### Fixed:
  - autoload in composer.json